feat: PageViewer backbone
Added in all the available add-ons as an array.

Passing in a reference of PageViewer into PageSelector and PageAdvancer so that we may know the state of any given page at any given time.

Starting extensive rework of PageViewer component so that we may interact with pages inside it using a unified interface.

Reworked all pages to forward a basic reference and crudely contain all the controls needed by the final version.